requirements.txt
pandas==2.2.3
numpy==2.1.3
pyarrow==18.1.0
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
holidays==0.62
tqdm==4.67.1

# Modeling
xgboost==2.1.3
scikit-learn==1.6.0

# Viz / app
plotly==5.24.1
streamlit==1.41.1

